+++ crates/ruff_python_ast/src/nodes.rs
@@ -1,11 +1,10 @@
 #![allow(clippy::derive_partial_eq_without_eq)]
 
 use itertools::Itertools;
+use malachite::Integer;
 use std::fmt;
 use std::fmt::Debug;
 use std::ops::Deref;
-
-use num_bigint::BigInt;
 
 use ruff_text_size::{Ranged, TextRange, TextSize};
 
@@ -2604,7 +2603,7 @@ pub enum Constant {
     Bool(bool),
     Str(StringConstant),
     Bytes(BytesConstant),
-    Int(BigInt),
+    Int(Integer),
     Float(f64),
     Complex { real: f64, imag: f64 },
     Ellipsis,
